made Listing and Map use the markets inside the Vuex store
